backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
openai
python-pptx
pymupdf
python-docx
langchain
sentence-transformers
python-dotenv
